pyproject.toml
[tool.poetry]
name = "coursework-7"
version = "0.1.0"
description = ""
authors = ["Anton Gromov <seed14112@yaandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
celery = "^5.4.0"
redis = "^5.1.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
